Fix output-neutral format/vararg defects (#7)
Format-string scan (issue #7, Phase 9) surfaced these byte-identical
fixes; the golden gate stays YES (debug + asan-ubsan) with or without
them.

- c2.c:446   times_masthead %*s field width is `67 - strlen(turn_s)`
             (size_t / unsigned long); a %*s width arg must be int. Cast
             (int) — representation-preserving for the small in-range
             value, and the LP64-correct fix (printf read 4 of 8 bytes
             of the size_t before).
- garr.c:427 vestigial add_s(n) extra arg on "%s pledges to us."
             (n was always 0, used nowhere else) → removed, and dropped
             the now-unused `int n`.
- main.c:656 sizeof(struct box) printed via %d (size_t) → %zu (startup
             stdout; not in the golden DB).
- stack.c:204/400 vestigial just_name() extra args on the VECT
             unstack/drop messages → removed.

+++ olympia/c2.c
@@ -443,7 +443,7 @@ times_masthead(void)
 	fprintf(fp, "   | The Olympia Times %50s |\n", issue_s);
 	fprintf(fp, "   | %-68s |\n", date);
 	fprintf(fp, "   |                                                                      |\n");
-	fprintf(fp, "   | %s %*s |\n", turn_s, 67 - strlen(turn_s), from_host);
+	fprintf(fp, "   | %s %*s |\n", turn_s, (int)(67 - strlen(turn_s)), from_host);
 	fprintf(fp, "   +----------------------------------------------------------------------+\n\n");
 	fprintf(fp, "     Questions, comments, to play: %s\n\n", game_url);
 	fprintf(fp, "                             Olympia PBEM\n\n");
